Add ability to use nested variables and initial data
It is now possible to provide nested dictionaries as variables and then
refer to them using a special variable syntax.
Example:

variables = {
  apache: {
    servername: 'testserver'
  }
}

If the above dictionary is fed into evaluate() and the path
"apache_servers/%{apache:servername}" is present in /etc/varstack.yaml
then this path would evaluate to "apache_servers/testserver.yaml".

evaluate() can now be called with a named parameter "init_data" that
should contain the initial dataset into which the data from the varstack
files is merged. Previously the initial data was always an empty
dictionary.